don't error if PR already exists
Ran into this when pushing out node 8 PRs.

This enables a `plan` + `push` iteration cycle where you iterate on
`plan`, and `push` the result. `push` either creates a new PR or
updates an existing PR.

Also tweaked `plan` to output the git diff if you give it a single
repo, which enables small-scale iteration on `plan`.