Explicitly meter calls to Github API. Remove retrier.
We do this because the Github API triggers abuse warnings if a single
user makes lots of concurrent requests. By having a shared ticker, we
ensure that Github API requests don't happen in parallel.

Moreover, we also ensure that there's a global rate limit of 1 request
to the GithubAPI per 720 milliseconds. We're following the approach
shown here: https://github.com/google/go-github/issues/431

This means that we'll stay within the rate limit requirement
of 5000 requests-per-hour for authenticated users.

```
5000 QPH = 83.3 QPM = 1.38 QPS = 720ms/query
```