allow init from file
Previously, we required the user to do a GitHub (or GitLab) search to
generate the initial list of repos.

What if you... already know... the repos you want to target? :)
The previous UX was actually more painful.

This should make it easier to just declare your repo list and move on.